+++ InputfieldURLChecker.js
@@ -1,135 +1,100 @@
 /**
  * InputfieldURLChecker module for ProcessWire
- * https://github.com/rolandtoth/InputfieldURLChecker
+ * https://goo.gl/F8CAK8
  */
 
 $(document).ready(function () {
 
     var IUC = {
         parent: '.InputfieldURL',
-        selector: 'input:not([type="hidden"])',
+        selector: 'a.iuc',
         asmSelectPlaceholder: 'data-asm-placeholder',
-        linkClass: 'iuc-button',
         linkHiddenClass: 'iuc-hide',
-        lockedFields: '.collapsed7, .collapsed8',
-        button: '<a class="iuc-button" target="_blank"><i class="fa fa-arrow-right"></i></a>',
+        lockedClass: 'iuc-locked-link',
         iframeSelector: 'iuc-iframe',
         overlaySelector: 'iuc-overlay',
         dummyFieldSelector: 'IUC-dummy',
-        mode: (config.IUC && config.IUC.mode) ? config.IUC.mode.toString().split(',') : 'button-left',
-        target: (config.IUC && config.IUC.target) ? config.IUC.target : 'new-window',
-        buttonPosition: (config.IUC && config.IUC.buttonPosition) ? config.IUC.buttonPosition : 'button-left',
-        forceHttp: (config.IUC && config.IUC.forceHttp !== 'undefined') ? config.IUC.forceHttp : true,
-        enabled_fields: (config.IUC && config.IUC.enabled_fields) ? config.IUC.enabled_fields : false,
-        enabled_templates: (config.IUC && config.IUC.enabled_templates) ? config.IUC.enabled_templates : false,
-        template: (config.IUC && config.IUC.template) ? config.IUC.template : false
+        dataMode: 'data-iuc-mode',
+        dataTarget: 'data-iuc-target',
+        dataForceHttp: 'data-iuc-force-http',
+        dataLoaded: 'data-iuc-loaded'
     };
-
-
-    // stop if template is not among enabled templates
-    // but skip if on own module edit page
-    if ($('form#ModuleEditForm').attr('action') !== 'edit?name=InputfieldURLChecker') {
-        if (IUC.enabled_templates && IUC.template) {
-            if (IUC.enabled_templates.indexOf(IUC.template) === -1) {
-                return true;
-            }
-        }
-    }
 
     // get button height with a dummy element
     $('body').append('<input class="' + IUC.dummyFieldSelector + '">');
     IUC.btnHeight = $('.' + IUC.dummyFieldSelector).outerHeight() - 2;
     $('.' + IUC.dummyFieldSelector).remove();
 
-    IUC.selector = IUC.parent + ' ' + IUC.selector;
-
     $(document).on('ready reloaded wiretabclick', initIUC);
 
     function initIUC() {
 
         // append empty iframe to the right on start
-        if (IUC.target === 'iframe' && !$('.' + IUC.iframeSelector).length) {
+        if ($('input[' + IUC.dataTarget + '="iframe"]').length && !$('.' + IUC.iframeSelector).length) {
             $('body').append('<iframe class="' + IUC.iframeSelector + '">');
         }
 
-        // linkify fields with locked status
-        if ($(IUC.parent).is(IUC.lockedFields)) {
+        $(IUC.selector).not('[' + IUC.dataLoaded + '="1"]').each(function () {
 
-            $(IUC.parent).filter(IUC.lockedFields).find('.InputfieldContent').each(function () {
+            if($(this).hasClass(IUC.lockedClass)) {
+                addLockedFieldMode($(this));
+            }
 
-                var content = $(this).html().trim();
+            var currInput = $(this).next('input:not([type="hidden"])');
 
-                if (content && content !== "" && content !== "&nbsp;" && !$(this).find('a').length) {
-                    $(this).wrapInner(function () {
-                        return '<a href="' + content + '" target="_blank"></a>';
+            // set link height
+            setTimeout(function () {
+                if (IUC.btnHeight > 0) {
+                    currInput.prev(IUC.selector).css({
+                        'height': IUC.btnHeight + 'px',
+                        //'line-height': IUC.btnHeight - 1 + 'px'
+                        'line-height': IUC.btnHeight + 'px'
                     });
                 }
-            });
+            }, 0);
 
-            return true;
-        }
+            addButtonMode(currInput);
 
-        else if ($(IUC.selector).length) {
+            $(this).attr(IUC.dataLoaded, 1);
+        });
 
-            $(IUC.selector).each(function () {
 
-                var currInput = $(this),
-                    fieldList;
+        $('input[' + IUC.dataMode + '!=""][' + IUC.dataMode + ']').not('[' + IUC.dataLoaded + '="1"]').each(function () {
 
-                // continue if field is not among enabled fields
-                if (IUC.enabled_fields) {
-                    fieldList = IUC.enabled_fields.replace(' ', '');
-                    fieldList = '#Inputfield_' + fieldList.replace(',', ', #Inputfield_');
+            var mode = $(this).attr(IUC.dataMode);
 
-                    if (!currInput.is(fieldList)) {
-                        return true;
-                    }
-                }
+            if (mode.indexOf('ctrl-shift-click') !== -1) {
+                addHotkeyMode($(this), 'click', 'ctrl-shift-click');
+            }
 
-                // add class to set padding only on enabled inputs
-                currInput.addClass('iuc-input');
+            if (mode.indexOf('ctrl-shift-enter') !== -1) {
+                addHotkeyMode($(this), 'keydown', 'ctrl-shift-enter');
+            }
 
-                // set link height
-                setTimeout(function () {
-                    if (IUC.btnHeight > 0) {
-                        currInput.next('.' + IUC.linkClass).css({
-                            'height': IUC.btnHeight + 'px',
-                            'line-height': IUC.btnHeight - 1 + 'px'
-                        });
-                    }
-                }, 0);
+            $(this).attr(IUC.dataLoaded, 1);
+        });
+    }
 
-                // disable adding link and events twice (ajax loaded tabs and fields)
-                if (currInput.next('.' + IUC.linkClass).length) {
-                    return true;
-                }
 
-                if (config.IUC.mode.indexOf('button') !== -1) {
-                    addButtonMode(currInput, IUC.buttonPosition);
-                }
+    function addLockedFieldMode(obj) {
 
-                if (config.IUC.mode.indexOf('ctrl-shift-click') !== -1) {
-                    addHotkeyMode('click', 'ctrl-shift-click');
-                }
+        obj.on('click', function (e) {
 
-                if (config.IUC.mode.indexOf('ctrl-shift-enter') !== -1) {
-                    addHotkeyMode('keydown', 'ctrl-shift-enter');
-                }
-            });
-        }
+            var url = $(this).attr('href');
+
+            e.preventDefault();
+
+            return url ? gotoLink(url, obj.attr(IUC.dataTarget)) : false;
+        });
     }
 
 
     function addButtonMode(obj, side) {
 
-        obj.after(IUC.button);
-
-        $('body').addClass('iuc-' + side);
-
         obj.on('keyup fieldchange', function () {
 
-            var url = getUrl(obj.val(), IUC.forceHttp),
-                link = obj.next('.' + IUC.linkClass);
+            var url = getUrl(obj.val(), $(this).attr(IUC.dataForceHttp)),
+                link = obj.prev(IUC.selector);
 
             link.attr('href', url);
 
@@ -137,13 +102,14 @@ $(document).ready(function () {
 
         });
 
-        obj.next('.' + IUC.linkClass).on('click', function (e) {
+        obj.prev(IUC.selector).on('click', function (e) {
 
-            var url = $(this).attr('href');
+            var url = $(this).attr('href'),
+                currInput = $(this).next('input:not([type="hidden"])');
 
             e.preventDefault();
 
-            return url ? gotoLink(url) : false;
+            return url ? gotoLink(url, currInput.attr(IUC.dataTarget)) : false;
         });
 
         // populate on load
@@ -153,21 +119,19 @@ $(document).ready(function () {
     }
 
 
-    function addHotkeyMode(eventName, mode) {
+    function addHotkeyMode(obj, eventName, mode) {
 
-        $(IUC.selector).on(eventName, function (e) {
+        obj.on(eventName, function (e) {
 
             var url = $(this).val(),
                 isCtrlShiftPressed = e.ctrlKey && e.shiftKey,
                 isEnterPressed = e.keyCode == 10 || e.keyCode == 13;
 
-            if (!url) {
-                return true;
-            }
+            if (!url) return true;
 
             if ((mode === 'ctrl-shift-enter' && isCtrlShiftPressed && isEnterPressed) ||
                 (mode === 'ctrl-shift-click' && isCtrlShiftPressed)) {
-                gotoLink(getUrl(url, IUC.forceHttp));
+                gotoLink(getUrl(url, $(this).attr(IUC.dataForceHttp)), $(this).attr(IUC.dataTarget));
                 return false;
             }
         });
@@ -181,11 +145,11 @@ $(document).ready(function () {
      * @param url
      * @returns {boolean}
      */
-    function gotoLink(url) {
+    function gotoLink(url, target) {
 
         var iframe = $('.' + IUC.iframeSelector);
 
-        if (IUC.target === 'new-window' || document.body.scrollWidth < 640) {
+        if (target === 'new-window' || document.body.scrollWidth < 640) {
             window.open(url);
 
         } else {
@@ -194,7 +158,7 @@ $(document).ready(function () {
             $('body').append('<div class="' + IUC.overlaySelector + '">');
             iframe.addClass('iuc-iframe-animate').attr('src', url);
 
-            iframe.on('load', function() {
+            iframe.on('load', function () {
                 $(this).addClass('iuc-iframe-loaded');
             })
         }
@@ -223,22 +187,21 @@ $(document).ready(function () {
 
     // remove overlay on click
     $(document).on('click', '.' + IUC.overlaySelector, function () {
-        IUCRemoveOverlay();
+        IUCCloseIframe();
     });
 
     // remove overlay on pressing ESC
     $(document).bind('keyup', '.' + IUC.overlaySelector, function (e) {
         if (e.keyCode === 27) {
-            IUCRemoveOverlay();
+            IUCCloseIframe();
         }
     });
 
-    function IUCRemoveOverlay() {
-        $('.' + IUC.iframeSelector).attr('src', '').removeClass('iuc-iframe-animate');
+    function IUCCloseIframe() {
+        $('.' + IUC.iframeSelector).removeClass('iuc-iframe-animate iuc-iframe-loaded').attr('src', '');
         $('html').css('overflow-y', '');
         $('.' + IUC.overlaySelector).remove();
     }
-
 });