+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 #Changelog
 
+### 1.1.0 (2015-11-22)
+
+- module was rewritten to use InputfieldURL::render hook instead JavaScript
+- fixed locked url fields when field description or notes were present
+- locked url fields respect module setting "target" to open url in new window or iframe
+
 ### 1.0.5 (2015-11-21)
 
 - fixed disfunctional "Enabled fields" (reported by Macrura)